refactor: extract LLM providers into WhisperDesk.Llm module
- Created src/llm/WhisperDesk.Llm.Contract (ILlmProvider, LlmRequestOptions)
- Created src/llm/providers/WhisperDesk.Llm.Provider.AzureOpenAI
- Created src/llm/WhisperDesk.Llm (DI assembly)
- Core now depends on Llm.Contract only, OpenAI package removed from Core
- Same pattern as STT module
