+++ WhisperDesk.sln
@@ -19,6 +19,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WhisperDesk.Stt.Provider.Az
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WhisperDesk.Stt.Provider.Volcengine", "src\stt\providers\WhisperDesk.Stt.Provider.Volcengine\WhisperDesk.Stt.Provider.Volcengine.csproj", "{484FB2A8-8EDC-4703-B93C-27A37085643F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WhisperDesk.Llm.Contract", "src\llm\WhisperDesk.Llm.Contract\WhisperDesk.Llm.Contract.csproj", "{6236DA27-5324-495F-A9CC-9C0179F628D7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WhisperDesk.Llm.Provider.AzureOpenAI", "src\llm\providers\WhisperDesk.Llm.Provider.AzureOpenAI\WhisperDesk.Llm.Provider.AzureOpenAI.csproj", "{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WhisperDesk.Llm", "src\llm\WhisperDesk.Llm\WhisperDesk.Llm.csproj", "{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -113,6 +119,42 @@ Global
 		{484FB2A8-8EDC-4703-B93C-27A37085643F}.Release|x64.Build.0 = Release|Any CPU
 		{484FB2A8-8EDC-4703-B93C-27A37085643F}.Release|x86.ActiveCfg = Release|Any CPU
 		{484FB2A8-8EDC-4703-B93C-27A37085643F}.Release|x86.Build.0 = Release|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Debug|x64.Build.0 = Debug|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Debug|x86.Build.0 = Debug|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Release|x64.ActiveCfg = Release|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Release|x64.Build.0 = Release|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Release|x86.ActiveCfg = Release|Any CPU
+		{6236DA27-5324-495F-A9CC-9C0179F628D7}.Release|x86.Build.0 = Release|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Debug|x64.Build.0 = Debug|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Debug|x86.Build.0 = Debug|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Release|x64.ActiveCfg = Release|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Release|x64.Build.0 = Release|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Release|x86.ActiveCfg = Release|Any CPU
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D}.Release|x86.Build.0 = Release|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Debug|x64.Build.0 = Debug|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Debug|x86.Build.0 = Debug|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Release|x64.ActiveCfg = Release|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Release|x64.Build.0 = Release|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Release|x86.ActiveCfg = Release|Any CPU
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -123,5 +165,8 @@ Global
 		{B65409DD-1615-4E5E-9E0B-A2876B80633A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{BE3CBEBE-D5AB-4323-86A9-B29595934512} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{484FB2A8-8EDC-4703-B93C-27A37085643F} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{6236DA27-5324-495F-A9CC-9C0179F628D7} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{95E822AD-4EFB-43A2-A347-EDE81D29BD2D} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{531EDA6D-BCC1-4C8E-8EF4-5B2507D401FE} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
