+++ src/WhisperDesk/App.xaml.cs
@@ -9,6 +9,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using WhisperDesk.Core.Configuration;
 using WhisperDesk.Stt;
+using WhisperDesk.Llm;
 using WhisperDesk.Core.Models;
 using WhisperDesk.Core.Pipeline;
 using WhisperDesk.Models;
@@ -126,6 +127,9 @@ public partial class App : Application
 
         // Register STT provider (from WhisperDesk.Stt)
         services.AddSttProvider(pipelineConfig.SttProvider, config);
+
+        // Register LLM provider (from WhisperDesk.Llm)
+        services.AddLlmProvider(pipelineConfig.LlmProvider, config);
 
         // Register Core pipeline services
         services.AddWhisperDeskPipeline(pipelineConfig, config);
