+++ src/llm/WhisperDesk.Llm.Contract/LlmRequestOptions.cs
@@ -1,0 +1,8 @@
+namespace WhisperDesk.Llm.Contract;
+
+public class LlmRequestOptions
+{
+    public float Temperature { get; init; } = 0.3f;
+    public int? MaxTokens { get; init; }
+    public IReadOnlyDictionary<string, object> ProviderOptions { get; init; } = new Dictionary<string, object>();
+}
