+++ src/llm/WhisperDesk.Llm.Contract/WhisperDesk.Llm.Contract.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <RootNamespace>WhisperDesk.Llm.Contract</RootNamespace>
+  </PropertyGroup>
+</Project>
