+++ src/llm/providers/WhisperDesk.Llm.Provider.AzureOpenAI/AzureOpenAILlmConfig.cs
@@ -1,9 +1,5 @@
-namespace WhisperDesk.Core.Providers.Llm.AzureOpenAI;
+namespace WhisperDesk.Llm.Provider.AzureOpenAI;
 
-/// <summary>
-/// Configuration for Azure OpenAI LLM provider.
-/// Bound from "AzureOpenAI" section in appsettings.json.
-/// </summary>
 public class AzureOpenAILlmConfig
 {
     public string Endpoint { get; set; } = string.Empty;
